scaffolding

+++ transactions/models.py
@@ -7,12 +7,12 @@ from accounts.models import ServiceProvider
 class Booking(models.Model):
     user = models.ForeignKey(User)
     service_provider = models.ForeignKey(ServiceProvider)
-    value = models.FloatField(blank=True, null=True)
+    value = models.DecimalField(max_digits=7, decimal_places=2, blank=True, null=True)
     currency = models.CharField(max_length=3, choices=transactions.CURRENCIES, default="eur")
     transaction_date = models.DateTimeField()
 
     def __str__(self):
-        return "%s %s: %s %f" % (self.user.first_name,
+        return "%s %s: %s %.2f" % (self.user.first_name,
                                  self.user.last_name,
                                  self.currency,
                                  self.value)
@@ -20,12 +20,12 @@ class Booking(models.Model):
 
 class Payment(models.Model):
     user = models.ForeignKey(User)
-    value = models.FloatField()
+    value = models.DecimalField(max_digits=7, decimal_places=2, )
     currency = models.CharField(max_length=3, choices=transactions.CURRENCIES, default='eur')
     transaction_date = models.DateTimeField()
 
     def __str__(self):
-        return "%s %f to %s %s at %s" % (self.currency,
+        return "%s %.2f to %s %s at %s" % (self.currency,
                                          self.value,
                                          self.user.first_name,
                                          self.user.last_name,
